Remove now-unused idle killing code
Now that the legacy trim code is gone (ag/26563712), mLowRamStartTime is
never set to a non-zero value and mLowRamTimeSinceLastIdle is never
updated. This means that doKilling in performIdleMaintenance is always
false, so the code guarded by that is all dead code.

Furthermore, the dead code was the only place that consumed
ProcessProfileRecord.mInitialIdlePssOrRss or
ProcessStateRecord.mNotCachedSinceIdle, so these can be removed as well.

Bug: 253914117
Test: build
Flag: EXEMPT refactor
Change-Id: Ib938d2cfe7fbb254a9fef3be11d0c3bcc20df15a