Hide lockscreen UI when transitioning to dream with shade open
When the dream starts, it requests to close all system dialogs. This
includes the notification shade. If the dream is started from
lockscreen while the shade is open, the shade quickly collapses then
the lockscreen UI is visible for a very short period of time before the
dream starts, which looks janky.

Instead, if the shade is open when this transition starts, we just keep
the lockscreen UI hidden the whole time.

Bug: 409720192
Test: atest LockscreenToDreamingTransitionViewModelTest
      (see bug for video)
Flag: NONE minor transition fix
Change-Id: I3d0529f647613e1e66acecb7075d114380fa3f4c