Attempt #3: Add smooth dimming transition for swapping in flexible split
Reason for revert: Debugged the issue. Test failure was due to a crash, caused by code not properly guarded by a flag check. Since test devices are not all at the same stage of flag rollout, and the crash resulted in confusing test failure messages, this was a hard one to figure out. Hopeful that this was the only thing causing the test failures.

CL description: Animates the dim layer during the swap movement, so that dimmed offscreen apps gradually become brighter as they come onscreen (and vice versa).

Previously, the shift was abrupt (at the end of the swap movement) and looked bad.

Also adds an API to SplitState to check the onscreen/offscreen status of a given Rect (should be useful later too).

Bug: 349828130
Flag: com.android.wm.shell.enable_flexible_two_app_split
Test: Dim layer transitions nicely when swapping. Doesn't affect onscreen > onscreen case. No change when flag is off.
Change-Id: Icdc0a8b7026aa36dcd5caf23d89c05f0da8edeb4