Specify windowing mode when starting single intent
Previously when starting a single intent we did not specify windowing
mode. This caused the task to keep its previous windowing mode. By
specifying the windowing mode, we enforce the launch to be in fullscreen
all the time.

Fix: 414808454
Test: Move a desktop app to split through app header, choose the same
app as the app pair. See that app is launched in fullscreen. App is a
single-instance app.
Flag: EXEMPT Bugfix (Enforcing policy of the method)

Change-Id: Ic6375dd2805350f15772e5e7338c7d1ced0cc6d7