Don't let windows in non-fill-screen windowing modes control insets
Controllable insets sources are attached to a display. If a window is in
a non-fill-screen windowing mode, (1) the window frame might not be able
to reach the insts sources, or (2) the window might need to share the
insets sources with others.

For (1), it makes less sense to let window to control insets sources,
since the window won't receive the insets regardless.

For (2), letting one of the windows control insets might make the layout
unstable when switching the control target to another window with a
different set of requested visible types.

This CL prevents windows in non-fill-screen windowing modes from
controlling insets.

Bug: 329124127
Flag: com.android.window.flags.force_show_system_bar_for_bubble
Test: atest InsetsPolicyTest
Change-Id: I97ae1040743f68a2b20cba1386a36ed4149f40d3