Only update latest surface visibility for transition participants
TransitionController#finishTransition invokes
ensureParticipantSurfaceVisibility, so when the playing transition
is finished, the surface visibility of participants will still be
updated.

This avoids setting intermediate state when a transition is finished
and its participant is collected by next transition, and then the
next transition checks it as a no-op change, so the condition
isPlayingTarget is false when the transition plays.

Bug: 416124146
Flag: com.android.window.flags.respect_hierarchy_surface_visibility
Test: Swipe to enter pip from a multi-activities task.
      The original task won't disappear a frame occasionally.
Change-Id: I0dc1a0fa33cf8c1867221d997b7fd297dfdec12e