Add parsing support for new purpose declaration manifest elements.
This is an attempt to reland commit 980c82db1bcf6d80bfffd13f64bb555097b5ab09 after it has been reverted due to perf regressions (boot-time). The hypothesis is that the regression was caused due to the population of new usesPermissionMapping that has been added in AndroidPackage. Considering there could be a lot of uses-permission use cases, this population of the map might have increased the latency. This CL flag guards adding entries into the map so the change can be relanded without any impact. The regression is expected to show up again when progressing the flag, but is reasonable to seek an override considering the parsed results are cached for subsequent boots.

Bug: 417787288
Bug: 420985780
Test: atest PermissionInfoTest AndroidPackageTest ParsedPermissionTest ParsedUsesPermissionTest
Test: atest PurposeDeclarationPermissionsTest
Flag: android.permission.flags.purpose_declaration_enabled
Change-Id: Ibbef57b8eb3a3ba1281656b6a1f6a73ca3137c18