Add parsing support for new purpose declaration manifest elements.
Implementation for parsing and processing the newly introduced manifest
elements for purpose declarations. Appropriate plumbing is done so that
PermissionInfo provides access to requiresPurpose and the set of valid
purposes of a permission. In addition, AndroidPackage has been updated
to include a new mapping of the requested permission name to the
ParsedUsesPermission object for quicker look-up of associated metadata
(e.g. when used to update permission policy).

Bug: 417787288
Test: atest PermissionInfoTest AndroidPackageTest ParsedPermissionTest ParsedUsesPermissionTest
Test: atest PurposeDeclarationPermissionsTest
Flag: android.permission.flags.purpose_declaration_enabled
Change-Id: I06f9c30f0e0c3c26077154001fff30fc02a492e6

Change-Id: If0ec90626971f1b6ff235d825125a09f49b4bdba