Fall back to small icon when AppIconProvider throws
The PromotedNotificationContentExtractor is passing in the SystemUI
context instead of the app context to fetch icons, which leads to
NameNotFoundException being thrown if the app is not installed in the
main profile. This is a temporary workaround to just fall back to the
small icon for AOD RONs when that happens.

Fix: 414830446
Flag: android.app.ui_rich_ongoing
Test: manual, install Notify only on work profile and post a RON; verify it doesn't crash
Change-Id: I30829c33196a807efe806f421418e654400c1e2b