Update DisplayLayouts upon deferred updates
We wanna make sure we also update the Shell-side
cache for display layouts when deferred display updates come in,
similar to how this is done with Transition triggered onDisplayChange
calls.

Also refactored DisplayChangeController a little to have
DisplayController injected into it, as per the past TODO in
DisplayController::ctor().

Bug: 416553608
Flag: EXEMPT bugfix
Test: atest DisplayChangeControllerTest
Test: manually fold/unfold while in PiP2 until PiP bounds are wrong.

Change-Id: I16f2144e947864cdb1eddd11072a4ab00250c6a7