Reduce the side button long press duration to 300ms
config_globalActionsKeyTimeout is the value used in SingleKeyGestureDetector for the long-press timeout. In Wear, this timeout is used as the duration a the side (aka STEM) button needs to be pressed to trigger its long-press behavior. We are reducing this duration to make the side button long more responsive across the platform.

Bug: 417288497
Test: N/A config change only
Flag: EXEMPT minor config change
Change-Id: I8c143de850b02104c06cd5da7359b6b0946e24bb