Reload statusbar icons correctly per-display
Before this change we were only reacting to configuration changes for the default display, and this was triggering the re-loading status bar icons in all displays regardless.

After this, we're reloading icon groups only for displays that had a configuration change.

Bug: 404242315
Test: StatusBarOrchestratorTest, StatusBarIconRefreshInteractorTest
Flag: com.android.systemui.shared.status_bar_connected_displays
Change-Id: Id6790212ca5c090b1702b80ec88e2e755db9109c