Refactor the timing to start LocalDataManager
In b/403752469, we submited a workaround to prevent crashes. The workaround starts LocalDataManager in the foreground when the AmbientVolumeUiController is created but it's better to do so in the background when the AmbientVolumeUiController.start() is called.

Prevent calling refresh() when LocalDataManager doesn't start yet in
loadDevices() and do the first refresh() once it starts.

Flag: EXEMPT bugfix
Bug: 417378166
Test: atest AmbientVolumeUiControllerTest
Test: atest HearingDevicesDialogDelegateTest
Change-Id: If6ce4ac11bb0a618876c940caa0e407641e80b8e