Fix app icon showing when grouped notif is dismissed via action
When a grouped notification was being dismissed as a result of pressing
on an action, the app icon would briefly become visible before the
notification disappeared. This was because we were resetting the
visibility set by NotificationGroupingUtil when the notification was
being removed from the group - instead, we should only do that when we
plan to keep the notification in the shade (a.k.a. when the removal is
actually just a move).

Flag: EXEMPT bugfix
Fix: 415665263
Test: NotificationChildrenContainerTest
Test: press "Done" on grouped notif and verify that the icon doesn't appear
Test: verify that the icon still comes back when a notification is no
longer grouped (e.g. when all other notifs in the group are gone)

Change-Id: I655c7f30ade4ee9b010f71dcb902afe8910f4ea4