WM tracing: use perfetto protos
Adapt the WindowManager tracing code to use the protos in
external/perfetto (SOT), instead of those in frameworks/base.

The protos in frameworks/base are a duplicate of the SOT that
exists to ease the transition to perfetto, but eventually
they are going to be deleted.

Flag: EXEMPT mechanical refactoring
Test: presubmit
Bug: 323165543
Change-Id: I4ab0a41e1241c7693a5d175b70c7202e986fb6c4