Use WindowRelayoutResult for all WM->Client layout
The interface between client and WM has multiple methods
passing around the same set of information. It turns out that
WindowRelayoutResult encapsulates this pretty well and is
parcelable which makes it work well as a binder `out` parameter.

Take the existing APIs which are passing layout information and
have them all use this one.

This includes pulling the SurfaceControl out of WindowRelayoutResult
since it is only used for relayout -- in other words, common stuff
stays in WindowRelayoutResult while each API's unique parameters
don't.

This reverts commit 0499d0a181ce0b7ffd158dd9e82bccebef22aa05.

Reason for revert: To re-apply after fixing robolectric shadow.

Bug: 385976595
Bug: 417782119
Change-Id: I842b98716369e7074f4ec8c7dc516d0644465328