Update child formatters for large clock with correct time format
Root cause was Ic522121ba8de7374caa0120e2c07ee2ca8e314ab which was
intended to dedupe Calendar and SimpleDateFormat. The formatters are
only partially shared and as a result a couple of additional objects
weren't correctly receiving the updated time format.

While it shouldn't be nessecary (as the calendar is fully shared) I'm
also defensively adding a similar update to TimeZone to ensure it's
correct when refreshing the time immediately after a change.

Bug: 417649342
Test: Manually checked large clock 24hr format
Flag: com.android.systemui.shared.clock_reactive_variants
Change-Id: Ieeb7825b71f3b56bc93841e8ffb18ea96e7e97e5