Avoid updating zombie provider's masked state
This change updates AppWidgetServiceImpl to
 - avoid updating the provider masked state if the provider is a zombie
 - reload provider masked state when the provider is reified

This fixes a bug where widgets are stuck in masked state after a
restore. This happens when the service masks a zombie provider in
reloadProviderMaskedStateLocked, but cannot unmask it in
updateWidgetPackageStoppedMaskedState because it cannot match the
PACKAGE_UNSTOPPED broadcast to the zombie provider's UID (because we do
not know the UID yet).

Bug: 372229859
Test: see bug for repro steps
Flag: EXEMPT bug fix
Change-Id: Ie4e8efe1d33890f6037dfd4074886f82652e45f2