Add tabs at the bottom of edit mode
The tabs change the grid header, as well as hide/show the available tiles.
I extracted common composables to avoid duplicating code, especially with the scrollable column and the available tiles grid. The code is otherwise unchanged for the current edit mode.

This is only the initial implementation and is missing key elements, such as:
- Final UI for the tabs bar
- Final icons for the tabs
- Showing/hiding badges
- Limiting resizing/reordering

Test: manually with and without the flag
Test: EditModeTest
Flag: com.android.systemui.qs_edit_mode_tabs
Bug: 416236871
Change-Id: I1f83082f4dd54fcb4528d1553953eefe3647521c