Don't consider re-ordering displays for transition root
When displays are just re-ordering, nothing about them
actually changes. However, these display re-orderings
usually happen along with other changes (since those
changes will cause displays to reorder). It Doesn't make
sense for the transition root to be at the display-level
in this case (and in-fact breaks a lot of assumptions) so
skip over re-order-only displays when searching for a
transition root.

Bug: 415274390
Test: atest TransitionTests#testCommonAncestor_excludeOrderOnlyDisplay
Flag: EXEMPT bugfix
Change-Id: I98175cd779af850cbef013451053a46217f99e11