No-op desk creation if display doesn't support desks.
Persistent repository defaults to restoring desks on default display. However, because createDesk does not include a check that the provided display supports desktop mode, this can result in non-desktop devices launching desks from overview.

This CL no-ops the creation in DesktopTasksController and DesktopRepositoryInitializerImpl in this particular case as a quick fix, but long term the plan is for the persistent repository to support restoring to external displays.

Bug: 417907552
Test: Manual; open desk on external display on phone, reboot device, confirm no desk appears in overview.
Flag: com.android.window.flags.enable_multiple_desktops_backend
Change-Id: I7c7acda85f130a1fc5176906efd145854d45abe3