Adding elevation to the expanded bubble handle view
As noted in the ticket, the BubbleBarHandleView remained present in the
BubbleBarExpandedView layout but was obscured by the TaskView (successor
of the SurfaceView). It appears that SurfaceView does not consistently
respect the view hierarchy, as bringing the handle view to the top had
no visible effect (even though the operation succeeded and the handle
view's order was correctly updated).
The only solution that resolved the issue was to add a tiny elevation,
to ensure that handle view is drawn on top of the task view.

Fixes: 416366725
Flag: EXEMPT - bug fix
Test: Manual. In unfolded state create any bubble. Fold the device and
expand just created bubble. Unfold the device.
Change-Id: Idc7d745091c01716ad83f29b3de23dee69d701ed