Allow moving WindowMagnification with a mouse even at the edge of a screen
When moving a window magnification with a mouse, it may be stack at the
edge of the screen, and a user have to perform drag move again to
actually magnify the edge of the screen.

With mouse input devices, even if the cursor is at the edge of the
screen, actual delta that the input device is emitting can be obtained.
We can use the value to move the magnification window so that users with
a mouse can smoothly magnify the entire screen.

Bug: 420776720
Test: MagnificationGestureDetectorTest
Flag: com.android.systemui.window_magnification_move_with_mouse_on_edge
Change-Id: I801f7994a9d360797d01b87df3199f8a14b67dec