[Bugfix]Execute screenTurnedOff after screenTurningOff is executed
The methods in screenTurningOff and screenTurnedOff are executed asynchronously, which causes mScreenOnEarly to be false occasionally when checkingIfReadyToSnapshot in screenTurningOff, resulting in snapshot failure. The startingWindow type after the screen is turned on is incorrectly set to splash screen.
Before the screenTurningOff method, mPendingScreenOffUnblocker is assigned in unBlockScreenOff, and mPendingScreenOffUnblocker is set to null after screenTurningOff is executed. Therefore, before executing screenTurnedOff, screenTurningOff is considered to have been executed and mPendingScreenOffUnblocker==null is judged.

Bug:414282309
Test: Manual
Flag: EXEMPT trivial change
Change-Id: I42a19655bbc16d3efa940bc981afb64c1d08dd91