BouncyBall: Have app log/exit if it loses foreground
Credit shayba@ for noting in an earlier code review that our tracking
of frame drops would be broken if the app lost focus.  At the time,
we put a line in the README to note we assume we're always in focus.

But with this CL, we put in code to have the app actively notice, log,
and (by default) exit in this situation.  This forces any automation
or user to realize there's a more fundamental issue than a frame
drop.

We also move our README to a markdown file, and add more information
to it.  We'll fill out the perfetto section later as we develop our
automation's use of perfetto to evaluate app runs.

Bug: 408044970
Test: Ran with and without ASSUMPTION_FAILURE_FORCES_EXIT, got the expected behaviors when having the app go out of focus.
Flag: EXEMPT for test app
Change-Id: Icd18ec1faa409670e901595b99eb8a03769af81e